Refactoring part 2

+++ peeplatex/backup.py
@@ -12,6 +12,7 @@ from datetime import datetime
 from operator import attrgetter
 
 from . import commands
+from . import fileio
 from ._shared import *
 
 
@@ -24,13 +25,11 @@ async def autosave():
     while True:
         try:
             await asyncio.sleep(interval)
-            # Right now we only have one article list at a time, and
-            #  changes just contains sentinel values.
             l = len(_g.changes)
             if _g.articleList and _g.currentPath and l != 0:
                 _debug(f"autosave: found {l} change{_p(l)}: "
                        f"{' '.join(_g.changes)}")
-                commands.write(silent=True)
+                fileio.write_articles(_g.articleList, _g.currentPath)
                 _debug("autosave complete")
                 _g.changes = []
         except asyncio.CancelledError:
@@ -49,13 +48,13 @@ def createBackup():
     if _g.articleList != [] and _g.currentPath is not None:
         dbName = _g.currentPath.name
         # Figure out the folder name
-        backupFolder = _g.currentPath.parent / "backups"
+        backupFolder = _g.currentPath / "backups"
         if not backupFolder.exists():
             backupFolder.mkdir()
         # Create the backup file
         now = datetime.now().strftime(".%y%m%d_%H%M%S")
         fname = backupFolder / (dbName + now)
-        commands.write(fname, silent=True)
+        fileio.write_articles(_g.articleList, backupFolder, (dbName + now))
         _debug("created backup file")
 
         # Create list of all backup files; most recent is last
@@ -63,7 +62,7 @@ def createBackup():
                          key=attrgetter('name'))
         # Check if the new backup is identical to the previous one. If so, we
         # delete the newest one instead of the oldest one.
-        if filecmp.cmp(backups[-1], backups[-2]):
+        if len(backups) >= 2 and filecmp.cmp(backups[-1], backups[-2]):
             # They are the same, delete the newest one
             backups[-1].unlink()
             backups.pop(-1)
